+++ src/sdk/account/toNexusAccount.addresses.test.ts
@@ -152,7 +152,7 @@ describe("nexus.account.addresses", async () => {
     const testnetAddress = await testnetClient.account.getAddress()
     const mainnetAddress = await mainnetClient.account.getAddress()
 
-    expect(testnetAddress).not.toBe(mainnetAddress)
+    expect(testnetAddress).toBe(mainnetAddress)
   })
 
   test("should test a mee account", async () => {