+++ src/sdk/modules/smartSessionsValidator/Helpers.ts
@@ -14,12 +14,9 @@ import {
   OWNABLE_VALIDATOR_ADDRESS,
   type PolicyData,
   REGISTRY_ADDRESS,
-  SMART_SESSIONS_ADDRESS,
-  type Session,
   encodeValidationData
 } from "../../constants"
 import { ERC7484RegistryAbi } from "../../constants/abi"
-import { SmartSessionAbi } from "../../constants/abi/SmartSessionAbi"
 import { parseReferenceValue } from "../utils/Helpers"
 import type { AnyData } from "../utils/Types"
 import type {
@@ -156,45 +153,6 @@ export const toActionConfig = (config: ActionConfig) => {
       })
     }
   }
-}
-
-/**
- * Gets the permission ID for a given session.
- *
- * @param client - The PublicClient to use for the contract call.
- * @param session - The Session object.
- * @returns A promise that resolves to the permission ID as a Hex string.
- */
-export const getPermissionId = async ({
-  client,
-  session
-}: {
-  client: PublicClient
-  session: Session
-}) => {
-  return (await client.readContract({
-    address: SMART_SESSIONS_ADDRESS,
-    abi: SmartSessionAbi,
-    functionName: "getPermissionId",
-    args: [session]
-  })) as Hex
-}
-
-export const isPermissionEnabled = async ({
-  client,
-  accountAddress,
-  permissionId
-}: {
-  client: PublicClient
-  accountAddress: Address
-  permissionId: Hex
-}) => {
-  return client.readContract({
-    address: SMART_SESSIONS_ADDRESS,
-    abi: SmartSessionAbi,
-    functionName: "isPermissionEnabled",
-    args: [permissionId, accountAddress]
-  })
 }
 
 /**